now gets posted to answer model
