+++ routes/api/quotes.js
@@ -2,11 +2,15 @@ const express = require('express');
 const Quote = require('../../models/Quote');
 // Quote model to be created
 const router = express.Router();
+const mailer = require('../../middleware/mailer')
 
 //Post /quotes (C)
 router.post('/', (req, res) => {
 	Quote.create(req.body)
-		.then((quote) => res.send(quote))
+		.then((quote) => {
+			mailer(req,res);
+			res.send(quote);
+		})
 		.catch((error) => res.sendStatus(500).json({ error: error.message }));
 });
