+++ views/contact.handlebars
@@ -24,7 +24,7 @@
       <div class="contact">
         <h3>Email Us</h3>
         {{msg}}
-        <form method="POST" actions="/post-quotes">
+        <form method="POST" action="/quotes">
           <p>
             <label>Name</label>
             <input type="text" name="name">
